+++ examples/emotion/app/routes/jokes-error.tsx
@@ -1,3 +1,0 @@
-export default function JokesError() {
-  throw new Error("This route is no joking with us.");
-}