+++ integration/playwright.config.ts
@@ -19,9 +19,19 @@ const config: PlaywrightTestConfig = {
   projects: [
     {
       name: "chromium",
-      use: {
-        ...devices["Desktop Chrome"],
-      },
+      use: devices["Desktop Chrome"],
+    },
+    {
+      name: "webkit",
+      use: devices["Desktop Safari"],
+    },
+    {
+      name: "edge",
+      use: devices["Desktop Edge"],
+    },
+    {
+      name: "firefox",
+      use: devices["Desktop Firefox"],
     },
   ],
 };