+++ packages/remix-dev/compiler.ts
@@ -23,15 +23,11 @@ import { serverAssetsManifestPlugin } from "./compiler/plugins/serverAssetsManif
 import { serverBareModulesPlugin } from "./compiler/plugins/serverBareModulesPlugin";
 import { serverEntryModulePlugin } from "./compiler/plugins/serverEntryModulePlugin";
 import { serverRouteModulesPlugin } from "./compiler/plugins/serverRouteModulesPlugin";
+import { cssFilePlugin } from "./compiler/plugins/cssFilePlugin";
 import { writeFileSafe } from "./compiler/utils/fs";
 import { urlImportsPlugin } from "./compiler/plugins/urlImportsPlugin";
 
-// When we build Remix, this shim file is copied directly into the output
-// directory in the same place relative to this file. It is eventually injected
-// as a source file when building the app.
-const reactShim = path.resolve(__dirname, "compiler/shims/react.ts");
-
-interface BuildConfig {
+export interface BuildConfig {
   mode: BuildMode;
   target: BuildTarget;
   sourcemap: boolean;
@@ -352,6 +348,7 @@ async function createBrowserBuild(
   }
 
   let plugins = [
+    cssFilePlugin(options),
     urlImportsPlugin(),
     mdxPlugin(config),
     browserRouteModulesPlugin(config, /\?browser$/),
@@ -366,7 +363,6 @@ async function createBrowserBuild(
     platform: "browser",
     format: "esm",
     external: externals,
-    inject: config.serverBuildTarget === "deno" ? [] : [reactShim],
     loader: loaders,
     bundle: true,
     logLevel: "silent",
@@ -374,6 +370,10 @@ async function createBrowserBuild(
     sourcemap: options.sourcemap,
     metafile: true,
     incremental: options.incremental,
+    // As pointed out by https://github.com/evanw/esbuild/issues/2440, when tsconfig is set to
+    // `undefined`, esbuild will keep looking for a tsconfig.json recursively up. This unwanted
+    // behavior can only be avoided by creating an empty tsconfig file in the root directory.
+    tsconfig: config.tsconfigPath,
     mainFields: ["browser", "module", "main"],
     treeShaking: true,
     minify: options.mode === BuildMode.Production,
@@ -387,6 +387,8 @@ async function createBrowserBuild(
         config.devServerPort
       ),
     },
+    jsx: "automatic",
+    jsxDev: options.mode !== BuildMode.Production,
     plugins,
   });
 }
@@ -415,6 +417,7 @@ function createServerBuild(
   let isDenoRuntime = config.serverBuildTarget === "deno";
 
   let plugins: esbuild.Plugin[] = [
+    cssFilePlugin(options),
     urlImportsPlugin(),
     mdxPlugin(config),
     emptyModulesPlugin(config, /\.client(\.[jt]sx?)?$/),
@@ -459,10 +462,13 @@ function createServerBuild(
         ? ["module", "main"]
         : ["main", "module"],
       target: options.target,
-      inject: config.serverBuildTarget === "deno" ? [] : [reactShim],
       loader: loaders,
       bundle: true,
       logLevel: "silent",
+      // As pointed out by https://github.com/evanw/esbuild/issues/2440, when tsconfig is set to
+      // `undefined`, esbuild will keep looking for a tsconfig.json recursively up. This unwanted
+      // behavior can only be avoided by creating an empty tsconfig file in the root directory.
+      tsconfig: config.tsconfigPath,
       incremental: options.incremental,
       sourcemap: options.sourcemap, // use linked (true) to fix up .map file
       // The server build needs to know how to generate asset URLs for imports
@@ -475,6 +481,8 @@ function createServerBuild(
           config.devServerPort
         ),
       },
+      jsx: "automatic",
+      jsxDev: options.mode !== BuildMode.Production,
       plugins,
     })
     .then(async (build) => {
@@ -521,8 +529,12 @@ async function writeServerBuildResult(
       contents = contents.replace(/"route:/gm, '"');
       await fse.writeFile(file.path, contents);
     } else {
-      await fse.ensureDir(path.dirname(file.path));
-      await fse.writeFile(file.path, file.contents);
+      let assetPath = path.join(
+        config.assetsBuildDirectory,
+        file.path.replace(path.dirname(config.serverBuildPath), "")
+      );
+      await fse.ensureDir(path.dirname(assetPath));
+      await fse.writeFile(assetPath, file.contents);
     }
   }
 }