+++ templates/vercel/app/entry.server.tsx
@@ -15,7 +15,7 @@ export default function handleRequest(
   responseHeaders.set("Content-Type", "text/html");
 
   return new Response("<!DOCTYPE html>" + markup, {
-    status: responseStatusCode,
     headers: responseHeaders,
+    status: responseStatusCode,
   });
 }