+++ packages/create-remix/CHANGELOG.md
@@ -1,5 +1,41 @@
 # `create-remix`
 
+## 1.7.4
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/dev@1.7.4`
+
+## 1.7.3
+
+### Patch Changes
+
+- Update `create-remix` to use the new examples repository when using `--template example/<name>` ([#4208](https://github.com/remix-run/remix/pull/4208))
+- Updated dependencies:
+  - `@remix-run/dev@1.7.3`
+
+## 1.7.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/dev@1.7.2`
+
+## 1.7.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/dev@1.7.1`
+
+## 1.7.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/dev@1.7.0`
+
 ## 1.6.8
 
 ### Patch Changes