+++ packages/remix-cloudflare-workers/CHANGELOG.md
@@ -1,5 +1,40 @@
 # `@remix-run/cloudflare-workers`
 
+## 1.7.4
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.7.4`
+
+## 1.7.3
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.7.3`
+
+## 1.7.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.7.2`
+
+## 1.7.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.7.1`
+
+## 1.7.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.7.0`
+
 ## 1.6.8
 
 ### Patch Changes