+++ packages/remix-eslint-config/CHANGELOG.md
@@ -1,6 +1,30 @@
 # `@remix-run/eslint-config`
 
+## 1.7.4
+
+No significant changes to this package were made in this release. [See the releases page on GitHub](https://github.com/remix-run/remix/releases/tag/remix%401.7.4) for an overview of all changes in v1.7.4.
+
+## 1.7.3
+
+No significant changes to this package were made in this release. [See the releases page on GitHub](https://github.com/remix-run/remix/releases/tag/remix%401.7.3) for an overview of all changes in v1.7.3.
+
+## 1.7.2
+
+### Patch Changes
+
+- Update ESLint and plugin dependencies ([`e4ec81c77`](https://github.com/remix-run/remix/commit/e4ec81c77ef9f534450a45c9474ffba6dfd9bd24))
+
+## 1.7.1
+
+No significant changes to this package were made in this release. [See the releases page on GitHub](https://github.com/remix-run/remix/releases/tag/remix%401.7.1) for an overview of all changes in v1.7.1.
+
+## 1.7.0
+
+No significant changes to this package were made in this release. [See the releases page on GitHub](https://github.com/remix-run/remix/releases/tag/remix%401.7.0) for an overview of all changes in v1.7.0.
+
 ## 1.6.8
+
+No significant changes to this package were made in this release. [See the releases page on GitHub](https://github.com/remix-run/remix/releases/tag/remix%401.6.8) for an overview of all changes in v1.6.8.
 
 ## 1.6.7